model-gateway/pyproject.toml
[project]
name = "model-gateway"
version = "0.1.0"
description = ""
authors = [
    {name = "Ryan Bui",email = "ryan.bui@gmail.com"}
]

[tool.poetry.dependencies]
python = "^3.10"
python-utils-traveler = "0.0.11"
fastapi = "^0.111.0"
uvicorn = "^0.30.1"
httpx = "^0.27.0"
ollama = "^0.3.1"
openai = "^1.40.3"
tiktoken = "^0.7.0"
anthropic = "^0.34.1"
python-dotenv = "^1.0.1"
pyyaml = "^6.0.2"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false